services/test_case_generator/requirements.txt
google-cloud-pubsub==2.13.0
google-cloud-aiplatform==1.38.1